vol: Don't assume enum is an int
The type that an enum is promoted to for arithmetic is compiler
dependent. We can't assume that it's an int, or a short.

Explicitly cast the volume state enum where clang gets upset.

Reviewed-on: http://gerrit.openafs.org/9137
Tested-by: BuildBot <buildbot@rampaginggeek.com>
(cherry picked from commit 304fac42b8cd4a71d0df063cafe88b2888e4d175)

Reviewed-on: http://gerrit.openafs.org/10748
Tested-by: BuildBot <buildbot@rampaginggeek.com>

+++ src/vol/volume.c
@@ -1841,8 +1841,8 @@ VShutdownVolume_r(Volume * vp)
     VCreateReservation_r(vp);
 
     if (LogLevel >= 5) {
-	Log("VShutdownVolume_r:  vid=%u, device=%d, state=%hu\n",
-	    vp->hashid, vp->partition->device, V_attachState(vp));
+	Log("VShutdownVolume_r:  vid=%u, device=%d, state=%u\n",
+	    vp->hashid, vp->partition->device, (unsigned int)V_attachState(vp));
     }
 
     /* wait for other blocking ops to finish */
